Mark give, reify et al. as NOINLINE on GHC 8.10 or older
This is needed to avoid
[GHC#16893](https://gitlab.haskell.org/ghc/ghc/issues/16893), a nasty
bug affecting GHC 8.10 and older in which `unsafeCoerce` can be
inlined too much, producing incorrect runtime results. #47
demonstrates a concrete example of `give` (which is defined in terms
of `unsafeCoerce`) actually going wrong at runtime, and since
`reify`/`reifyNat`/`reifySymbol` are also defined directly in terms
of `unsafeCoerce`, they also run the same risk.

This works around the problem by marking `give`, `reify`, etc. as
`NOINLINE` on GHC 8.10 or older to prevent GHC#16893 from happening.
The code from #47 has been converted to a regression test to avoid
this sort of thing from regressing in the future.

Fixes #47.