Use new, safer GHC.TypeLits API to define reify{Nat,Symbol}
With `base-4.18` (GHC 9.6), we can now define `reifyNat` and `reifySymbol`
without any uses of `unsafeCoerce`. Hooray! What's more, this does not exhibit
the incoherence issues seen when trying to use `withDict` to define `reify` or
`give`. See this comment for an explanation of why this is the case:
https://github.com/haskell/core-libraries-committee/issues/85#issuecomment-1229601325

This addresses one part of #44. It does not fix it completely, as it does not
address `reify` and `give`, which are separate matters.